requirements.txt
flask==2.3.0
python-dotenv==1.0.0
translate==3.8.0
requests==2.32.5

